+++ docs-markdown/markdownlint-custom-rules/rules.js
@@ -1,0 +1,10 @@
+// @ts-check
+
+"use strict";
+
+module.exports = [
+  require("./syntax"),
+  require("./zone"),
+  require("./moniker"),
+  require("./securelinks")
+];