+++ docs-markdown/src/extension.ts
@@ -6,7 +6,7 @@
  Logging, Error Handling, VS Code window updates, etc.
 */
 
-import * as vscode from "vscode";
+import { commands, ConfigurationTarget, ExtensionContext, window, workspace } from "vscode";
 import { insertAlertCommand } from "./controllers/alert-controller";
 import { boldFormattingCommand } from "./controllers/bold-controller";
 import { codeFormattingCommand } from "./controllers/code-controller";
@@ -20,12 +20,11 @@ import { quickPickMenuCommand } from "./controllers/quick-pick-menu-controller";
 import { insertSnippetCommand } from "./controllers/snippet-controller";
 import { insertTableCommand } from "./controllers/table-controller";
 import { checkExtension, generateTimestamp } from "./helper/common";
-import * as log from "./helper/log";
 import { UiHelper } from "./helper/ui";
 import { Reporter } from "./telemetry/telemetry";
 
-export const output = vscode.window.createOutputChannel("docs-markdown");
-export const masterRedirectOutput = vscode.window.createOutputChannel("docs-markdown-master-redirect");
+export const output = window.createOutputChannel("docs-markdown");
+export const masterRedirectOutput = window.createOutputChannel("docs-markdown-master-redirect");
 
 /**
  * Provides the commands to the extension. This method is called when extension is activated.
@@ -35,14 +34,18 @@ export const masterRedirectOutput = vscode.window.createOutputChannel("docs-mark
  *
  * param {vscode.ExtensionContext} the context the extension runs in, provided by vscode on activation of the extension.
  */
-export function activate(context: vscode.ExtensionContext) {
-    log.debug("Activating Markdown Authoring Extension.");
+export function activate(context: ExtensionContext) {
+    const { msTimeValue } = generateTimestamp();
+    output.appendLine(`[${msTimeValue}] - Activating docs markdown extension.`);
 
     // Places "Docs Markdown Authoring" into the Toolbar
     new UiHelper().LoadToolbar();
 
     // check for docs extensions
     installedExtensionsCheck();
+
+    // Markdownlint custom rule check
+    checkMarkdownlintCustomProperty();
 
     // Creates an array of commands from each command file.
     const AuthoringCommands: any = [];
@@ -66,24 +69,24 @@ export function activate(context: vscode.ExtensionContext) {
     try {
         AuthoringCommands.map((cmd: any) => {
             const commandName = cmd.command;
-            const command = vscode.commands.registerCommand(commandName, cmd.callback);
+            const command = commands.registerCommand(commandName, cmd.callback);
             context.subscriptions.push(command);
         });
     } catch (error) {
-        log.error("Error registering commands with vscode extension context: " + error);
+        output.appendLine(`[${msTimeValue}] - Error registering commands with vscode extension context: + ${error}`);
     }
 
-    log.debug("Registered commands with vscode extension context.");
+    output.appendLine(`[${msTimeValue}] - Registered commands with vscode extension context.`);
 
     // if the user changes markdown.showToolbar in settings.json, display message telling them to reload.
-    vscode.workspace.onDidChangeConfiguration((e: any) => {
+    workspace.onDidChangeConfiguration((e: any) => {
 
         if (e.affectsConfiguration("markdown.showToolbar")) {
 
-            vscode.window.showInformationMessage("Your updated configuration has been recorded, but you must reload to see its effects.", "Reload")
+            window.showInformationMessage("Your updated configuration has been recorded, but you must reload to see its effects.", "Reload")
                 .then((res) => {
                     if (res === "Reload") {
-                        vscode.commands.executeCommand("workbench.action.reloadWindow");
+                        commands.executeCommand("workbench.action.reloadWindow");
                     }
                 });
         }
@@ -91,17 +94,57 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 export function installedExtensionsCheck() {
+    const { msTimeValue } = generateTimestamp();
     // create a list to house docs extension names, loop through
     const docsExtensions = [
         "docsmsft.docs-article-templates",
         "docsmsft.docs-preview",
     ];
-    const { msTimeValue } = generateTimestamp();
     docsExtensions.forEach((extensionName) => {
         const friendlyName = extensionName.split(".").reverse()[0];
         const inactiveMessage = `[${msTimeValue}] - The ${friendlyName} extension is not installed.`;
         checkExtension(extensionName, inactiveMessage);
     });
+}
+
+/**
+ * Method to check for the docs custom markdownlint value.
+ * Checks for markdownlint.customRules property.  If markdownlint isn't installed, do nothing.  If markdownlint is installed, check for custom property values.
+ */
+export function checkMarkdownlintCustomProperty() {
+    const { msTimeValue } = generateTimestamp();
+    const customProperty = "markdownlint.customRules";
+    const customRuleset = "{docsmsft.docs-markdown}/markdownlint-custom-rules/rules.js";
+    const customPropertyData = workspace.getConfiguration().inspect(customProperty);
+    // new list for string comparison and updating.
+    const existingUserSettings: string[] = [];
+    if (customPropertyData) {
+        // if the markdownlint.customRules property exists, pull the global values (user settings) into a string.
+        if (customPropertyData.globalValue) {
+            const valuesToString = customPropertyData.globalValue.toString();
+            const individualValues = valuesToString.split(",");
+            individualValues.forEach((setting) => {
+                existingUserSettings.push(setting);
+            });
+            // if the customRuleset already exist, write a notification to the output window and continue.
+            if (existingUserSettings.indexOf(customRuleset) > -1) {
+                output.appendLine(`[${msTimeValue}] - Docs custom markdownlint ruleset is already set at a global level.`);
+            } else {
+                // if the customRuleset does not exists, append it to the other values in the list if there are any or add it as the only value.
+                existingUserSettings.push(customRuleset);
+                // update the user settings with new/updated values and notify user.
+                // if a user has specific workspace settings for customRules, vscode will use those. this is done so we don't override non-docs repos.
+                workspace.getConfiguration().update(customProperty, existingUserSettings, ConfigurationTarget.Global);
+                output.appendLine(`[${msTimeValue}] - Docs custom markdownlint ruleset added to user settings.`);
+            }
+        }
+        // if no custom rules exist, create array and add docs custom ruleset.
+        if (customPropertyData.globalValue === undefined) {
+            const customPropertyValue = [customRuleset];
+            workspace.getConfiguration().update(customProperty, customPropertyValue, ConfigurationTarget.Global);
+            output.appendLine(`[${msTimeValue}] - Docs custom markdownlint ruleset added to user settings.`);
+        }
+    }
 }
 
 // this method is called when your extension is deactivated