+++ docs-article-templates/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.1.4 (December 3rd, 2018)
+
+- Update command palette
+
+## 0.1.3 (October 3rd, 2018)
+
+- Remove vscode-extension-telemetry dependency
+- Update command name
+
+## 0.1.2 (June 25th, 2018)
+
+- Performance improvements
+
 ## 0.1.1 (June 20th, 2018)
 
 - Update readme links